refactor line num to a strategy
