+++ src/codemodder/dependency_management/base_dependency_writer.py
@@ -2,9 +2,10 @@ from abc import ABCMeta, abstractmethod
 from pathlib import Path
 from typing import Optional
 from codemodder.project_analysis.file_parsers.package_store import PackageStore
-from codemodder.change import ChangeSet
+from codemodder.change import Action, Change, ChangeSet, PackageAction, Result
 from codemodder.dependency import Dependency
 from packaging.requirements import Requirement
+from typing import Callable, Union, List
 
 
 class DependencyWriter(metaclass=ABCMeta):
@@ -38,3 +39,25 @@ class DependencyWriter(metaclass=ABCMeta):
                 self.dependency_store.dependencies.add(requirement)
                 new.append(new_dep)
         return new
+
+    def build_changes(
+        self,
+        dependencies: list[Dependency],
+        line_number_strategy: Callable,
+        strategy_arg: Union[int, List[str], List[int]],
+    ) -> list[Change]:
+        return [
+            Change(
+                lineNumber=line_number_strategy(strategy_arg, i),
+                description=dep.build_description(),
+                # Contextual comments should be added to the right side of split diffs
+                properties={
+                    "contextual_description": True,
+                    "contextual_description_position": "right",
+                },
+                packageActions=[
+                    PackageAction(Action.ADD, Result.COMPLETED, str(dep.requirement))
+                ],
+            )
+            for i, dep in enumerate(dependencies)
+        ]
