+++ src/codemodder/dependency_management/pyproject_writer.py
@@ -2,10 +2,14 @@ import tomlkit
 from typing import Optional
 from copy import deepcopy
 from codemodder.dependency import Dependency
-from codemodder.change import Action, Change, ChangeSet, PackageAction, Result
+from codemodder.change import ChangeSet
 from codemodder.dependency_management.base_dependency_writer import DependencyWriter
 from codemodder.diff import create_diff_and_linenums
 from codemodder.logging import logger
+
+
+def added_line_nums_strategy(lines, i):
+    return lines[i]
 
 
 class PyprojectWriter(DependencyWriter):
@@ -31,31 +35,14 @@ class PyprojectWriter(DependencyWriter):
             with open(self.path, "w", encoding="utf-8") as f:
                 tomlkit.dump(pyproject, f)
 
-        changes = self.build_changes(dependencies, added_line_nums)
+        changes = self.build_changes(
+            dependencies, added_line_nums_strategy, added_line_nums
+        )
         return ChangeSet(
             str(self.path.relative_to(self.parent_directory)),
             diff,
             changes=changes,
         )
-
-    def build_changes(
-        self, dependencies: list[Dependency], added_line_nums: list[int]
-    ) -> list[Change]:
-        return [
-            Change(
-                lineNumber=added_line_nums[i],
-                description=dep.build_description(),
-                # Contextual comments should be added to the right side of split diffs
-                properties={
-                    "contextual_description": True,
-                    "contextual_description_position": "right",
-                },
-                packageActions=[
-                    PackageAction(Action.ADD, Result.COMPLETED, str(dep.requirement))
-                ],
-            )
-            for i, dep in enumerate(dependencies)
-        ]
 
     def _parse_file(self):
         with open(self.path, encoding="utf-8") as f:
