+++ src/codemodder/dependency_management/requirements_txt_writer.py
@@ -1,8 +1,12 @@
 from typing import Optional
 from codemodder.dependency_management.base_dependency_writer import DependencyWriter
-from codemodder.change import Action, Change, ChangeSet, PackageAction, Result
+from codemodder.change import ChangeSet
 from codemodder.diff import create_diff
 from codemodder.dependency import Dependency
+
+
+def original_lines_strategy(original_lines, i):
+    return len(original_lines) + i + 1
 
 
 class RequirementsTxtWriter(DependencyWriter):
@@ -23,31 +27,14 @@ class RequirementsTxtWriter(DependencyWriter):
             with open(self.path, "w", encoding="utf-8") as f:
                 f.writelines(updated_lines)
 
-        changes = self.build_changes(dependencies, original_lines)
+        changes = self.build_changes(
+            dependencies, original_lines_strategy, original_lines
+        )
         return ChangeSet(
             str(self.path.relative_to(self.parent_directory)),
             diff,
             changes=changes,
         )
-
-    def build_changes(
-        self, dependencies: list[Dependency], original_lines: list[str]
-    ) -> list[Change]:
-        return [
-            Change(
-                lineNumber=len(original_lines) + i + 1,
-                description=dep.build_description(),
-                # Contextual comments should be added to the right side of split diffs
-                properties={
-                    "contextual_description": True,
-                    "contextual_description_position": "right",
-                },
-                packageActions=[
-                    PackageAction(Action.ADD, Result.COMPLETED, str(dep.requirement))
-                ],
-            )
-            for i, dep in enumerate(dependencies)
-        ]
 
     def _parse_file(self):
         with open(self.path, "r", encoding="utf-8") as f:
